\name{runShinyPIPE}
\alias{runShinyPIPE}

\title{
A Shiny GUI implementation of the PIPE design
}
\description{
Runs a Shiny web browser interface of the Product of Independent beta Probabilities dose Escalation (PIPE) design}
\usage{
runShinyPIPE()
}



\details{
This function implements a GUI web browser version of pipe.design, thus allowing easy usage and communication of the design to researchers.
}

\references{
 Mander A.P., Sweeting M.J. A product of independent beta probabilities dose escalation design for dual-agent phase I trials. \emph{Statistics in Medicine} (2015) 34(8): 1261--1276.
 
}

\author{
Chris Habron (Roche) and 
Michael Sweeting \email{mjs212@medschl.cam.ac.uk} (University of Cambridge, UK) 

}

\seealso{
\code{\link{pipe.design}}
}





